Fixed the casting error while putting a new csp violation report

+++ nanopay/src/net/nanopay/security/csp/CSPViolation.js
@@ -6,7 +6,7 @@ foam.CLASS({
 
   properties: [
     {
-      class: 'Int',
+      class: 'Long',
       name: 'id',
       doumentation: 'Sequence number.'
     },
